media/gstreamer: Attempt to use GL for video for Windows
Add a stab to support sharing the WGL context in GDK with the WGL
context in GStreamer, so that we can also use OpenGL in the gstreamer
media backend to playback videos.

Due to an issue in libepoxy, we can't just use
gdk_gl_context_make_current() from the media backend, which will cause a
crash as we attempt to make calls accross threads.  Instead, we need to
pull out the HGLRC (WGL context) from the GdkGLContext using the private
API we added in the previous commit, and call the native
wglMakeCurrent() in the system-provided opengl32.dll implementation.  We
also need to deactivate the GL context so that the wglShareLists() call
in libgstgl won't fail as we are only allowed to have a WGL context
current per thread.

This is currently WIP as playback of audio files work, but playback of
videos result in a shader compilation failure in libgstgl.